refactor: Changes the helper to a VariantImport component functioning as an extra layer on top of the UploadButton with specific logic for variants
